backend-old/requirements.txt
fastapi
uvicorn
celery
redis
anthropic
python-dotenv
sse-starlette
typing-extensions
pydantic
pydantic-settings
asyncio
groq
google-genai
pillow
httpx
modal
torch
numpy